+++ test/all_tests.dart
@@ -5,7 +5,7 @@
 
 library constrain.all.test;
 
-import 'package:unittest/unittest.dart';
+import 'package:test/test.dart';
 import 'shelf_auth_test.dart' as core;
 import 'authenticators/all_authenticator_tests.dart' as authenticators;
 import 'session/jwt/all_jwt_session_tests.dart' as jwt;