+++ test/authorisers/authorisation_test.dart
@@ -8,7 +8,7 @@ library shelf_auth.authorisation.test;
 import 'package:shelf_auth/shelf_auth.dart';
 import 'package:shelf_auth/src/authentication_impl.dart';
 
-import 'package:unittest/unittest.dart';
+import 'package:test/test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:shelf/shelf.dart';
 import 'dart:async';