+++ test/session/jwt/all_jwt_session_tests.dart
@@ -5,7 +5,7 @@
 
 library shelf_auth.session.jwt.all.test;
 
-import 'package:unittest/unittest.dart';
+import 'package:test/test.dart';
 import 'jwt_session_test.dart' as jwt_session;
 import 'jwt_session_auth_test.dart' as auth;
 import 'jwt_session_handler_test.dart' as handler;