+++ test/session/jwt/jwt_session_auth_test.dart
@@ -8,7 +8,7 @@ library shelf_auth.authentication.session.jwt.test;
 import 'package:shelf/shelf.dart';
 import 'dart:async';
 import 'package:option/option.dart';
-import 'package:unittest/unittest.dart';
+import 'package:test/test.dart';
 import 'package:shelf_auth/src/principal/user_lookup.dart';
 import 'package:shelf_auth/src/session/jwt/jwt_session_auth.dart';
 import 'package:shelf_auth/src/session/jwt/jwt_session.dart';