+++ test/session/jwt/jwt_session_test.dart
@@ -6,7 +6,7 @@
 library shelf_auth.session.jwt.test;
 
 import 'package:shelf_auth/src/session/jwt/jwt_session.dart';
-import 'package:unittest/unittest.dart';
+import 'package:test/test.dart';
 import 'package:dart_jwt/dart_jwt.dart';
 
 main() {